+++ modules/monitoring_function/r-function.tf
@@ -29,7 +29,9 @@ module "function" {
   function_app_version        = 3
 
   # Allow the current run to access Storage Table for FAME queries
-  storage_account_authorized_ips = ["${data.http.myip.body}/32"]
+  storage_account_network_rules_enabled = var.storage_account_network_rules_enabled
+  storage_account_network_bypass        = var.storage_account_network_bypass
+  storage_account_authorized_ips        = coalescelist(var.storage_account_authorized_ips, ["${data.http.myip.body}/32"])
 
   application_zip_package_path = var.zip_package_path